NSW calibration - Optimize memory footprint of conditions data
NSW calibration - Optimize memory footprint of conditions data

+++ MuonSpectrometer/MuonCalib/NSWCalib/NSWCalibTools/src/NSWCalibTool.h
@@ -44,24 +44,24 @@ namespace Muon {
     StatusCode calibrateStrip(const EventContext& ctx, const Muon::STGC_RawData* sTGCRawData, NSWCalib::CalibratedStrip& calibStrip) const override;
     
     
-    bool tdoToTime  (const EventContext& ctx, const bool inCounts, const int tdo, const Identifier& chnlId, double& time, const int relBCID) const override;
-    bool timeToTdo  (const EventContext& ctx, const double time, const Identifier& chnlId, int& tdo, int& relBCID) const override;
-    bool chargeToPdo(const EventContext& ctx, const double charge, const Identifier& chnlId, int& pdo) const override;
-    bool pdoToCharge(const EventContext& ctx, const bool inCounts, const int pdo, const Identifier& chnlId, double& charge) const override;
+    bool tdoToTime  (const EventContext& ctx, const bool inCounts, const int tdo, const Identifier& chnlId, float& time, const int relBCID) const override;
+    bool timeToTdo  (const EventContext& ctx, const float time, const Identifier& chnlId, int& tdo, int& relBCID) const override;
+    bool chargeToPdo(const EventContext& ctx, const float charge, const Identifier& chnlId, int& pdo) const override;
+    bool pdoToCharge(const EventContext& ctx, const bool inCounts, const int pdo, const Identifier& chnlId, float& charge) const override;
 
     virtual StatusCode initialize() override;  
 
     NSWCalib::MicroMegaGas mmGasProperties() const override;
 
-    inline double mmPeakTime  () const override {return m_mmPeakTime;  }
-    inline double stgcPeakTime() const override {return m_stgcPeakTime;}
+    inline float mmPeakTime  () const override {return m_mmPeakTime;  }
+    inline float stgcPeakTime() const override {return m_stgcPeakTime;}
 
   private:
     const NswCalibDbTimeChargeData* getCalibData(const EventContext& ctx) const;
     bool loadMagneticField(const EventContext& ctx, MagField::AtlasFieldCache& fieldCache ) const;
 
-    bool timeToTdoMM(const NswCalibDbTimeChargeData* tdoPdoData, const double time, const Identifier& chnlId, int& tdo, int& relBCID) const;
-    bool timeToTdoSTGC(const NswCalibDbTimeChargeData* tdoPdoData, const double time, const Identifier& chnlId, int& tdo, int& relBCID) const;
+    bool timeToTdoMM(const NswCalibDbTimeChargeData* tdoPdoData, const float time, const Identifier& chnlId, int& tdo, int& relBCID) const;
+    bool timeToTdoSTGC(const NswCalibDbTimeChargeData* tdoPdoData, const float time, const Identifier& chnlId, int& tdo, int& relBCID) const;
 
     ServiceHandle<Muon::IMuonIdHelperSvc> m_idHelperSvc {this, "MuonIdHelperSvc", "Muon::MuonIdHelperSvc/MuonIdHelperSvc"};
     SG::ReadCondHandleKey<AtlasFieldCacheCondObj> m_fieldCondObjInputKey {this, "AtlasFieldCacheCondObj", "fieldCondObj"};
@@ -77,17 +77,17 @@ namespace Muon {
     Gaudi::Property<double> m_longDiff{this, "longDiff", 0.019}; // mm/ mm
     Gaudi::Property<double> m_transDiff{this, "transDiff", 0.036};
     Gaudi::Property<double> m_ionUncertainty{this,"ionUncertainty", 4.0}; //ns
-    Gaudi::Property<double> m_mmPeakTime{this, "mmPeakTime", 200.}; //ns
-    Gaudi::Property<double> m_stgcPeakTime{this, "sTgcPeakTime", 0.}; // ns
+    Gaudi::Property<float> m_mmPeakTime{this, "mmPeakTime", 200.}; //ns
+    Gaudi::Property<float> m_stgcPeakTime{this, "sTgcPeakTime", 0.}; // ns
     Gaudi::Property<std::string> m_gasMixture{this, "GasMixture",  "ArCo2_937"};
 
     // these values should go into the conditions database, but introduce them as properties for now
     // they describe the shift of the time to get it into the VMM time window
-    Gaudi::Property<double> m_mmLatencyMC{this,"mmLatencyMC",25};
-    Gaudi::Property<double> m_mmLatencyData{this,"mmLatencyData",50}; //this is temporary, need to align with what we find in data
+    Gaudi::Property<float> m_mmLatencyMC{this,"mmLatencyMC",25};
+    Gaudi::Property<float> m_mmLatencyData{this,"mmLatencyData",50}; //this is temporary, need to align with what we find in data
     
-    Gaudi::Property<double> m_stgcLatencyMC{this,"stgcLatencyMC",-50};
-    Gaudi::Property<double> m_stgcLatencyData{this,"stgcLatencyData",-50}; //this is temporary, need to align with what we find in data
+    Gaudi::Property<float> m_stgcLatencyMC{this,"stgcLatencyMC",-50};
+    Gaudi::Property<float> m_stgcLatencyData{this,"stgcLatencyData",-50}; //this is temporary, need to align with what we find in data
 
     double m_interactionDensitySigma{0.0F};
     double m_interactionDensityMean{0.0F};
